feat: check related user when destroy departments

+++ app/Http/Controllers/Pc/DepartmentController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Pc\DepartmentRequest;
 use App\Http\Resources\Pc\DepartmentResource;
 use App\Models\Department;
+use App\Models\User;
 use Illuminate\Http\Response;
 
 class DepartmentController extends Controller
@@ -34,6 +35,10 @@ class DepartmentController extends Controller
 
     public function destroy(Department $department)
     {
+        $descendantIds = Department::descendantsAndSelf($department->id)->pluck('id');
+        if(User::whereIn('id', $descendantIds)->exists()){
+            return error("单位或其下级单位已关联人员，请先解除关联", Response::HTTP_UNPROCESSABLE_ENTITY);
+        }
         $department->delete();
         return no_content();
     }
